+++ src/lib/pages/account-details/types.ts
@@ -20,24 +20,18 @@ export interface NonRedelegatable {
 }
 
 export interface UserDelegationsData {
+  isLoading: boolean;
   stakingParams: Option<StakingParams>;
   isValidator: Option<boolean>;
-  isLoading: Option<boolean>;
   totalBonded: Option<Record<string, TokenWithValue>>;
-  isLoadingTotalBonded: boolean;
   totalDelegations: Option<Record<string, TokenWithValue>>;
   delegations: Option<Delegation[]>;
-  isLoadingDelegations: boolean;
   totalUnbondings: Option<Record<string, TokenWithValue>>;
   unbondings: Option<Unbonding[]>;
-  isLoadingUnbondings: boolean;
   totalRewards: Option<Record<string, TokenWithValue>>;
   rewards: Option<Record<string, TokenWithValue[]>>;
-  isLoadingRewards: boolean;
   redelegations: Option<Redelegation[]>;
-  isLoadingRedelegations: boolean;
   totalCommission: Option<Record<string, TokenWithValue>>;
-  isLoadingTotalCommission: boolean;
 }
 
 export enum TabIndex {